// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ThemeMode = {
  LIGHT: 'LIGHT',
  DARK: 'DARK'
} as const

export type ThemeMode = (typeof ThemeMode)[keyof typeof ThemeMode]


export const EmailProviders = {
  GMAIL: 'GMAIL',
  OUTLOOK: 'OUTLOOK',
  YAHOO: 'YAHOO',
  CUSTOM: 'CUSTOM'
} as const

export type EmailProviders = (typeof EmailProviders)[keyof typeof EmailProviders]


export const ImapAuthType = {
  OAUTH: 'OAUTH',
  PASSWORD: 'PASSWORD'
} as const

export type ImapAuthType = (typeof ImapAuthType)[keyof typeof ImapAuthType]


export const SmtpAuthType = {
  OAUTH: 'OAUTH',
  PASSWORD: 'PASSWORD'
} as const

export type SmtpAuthType = (typeof SmtpAuthType)[keyof typeof SmtpAuthType]


export const FolderType = {
  INBOX: 'INBOX',
  SENT: 'SENT',
  SPAM: 'SPAM',
  PHISHING: 'PHISHING',
  TRASH: 'TRASH',
  CUSTOM: 'CUSTOM',
  MALWARE: 'MALWARE'
} as const

export type FolderType = (typeof FolderType)[keyof typeof FolderType]


export const EmailFlagType = {
  SEEN: 'SEEN',
  FLAGGED: 'FLAGGED',
  ANSWERED: 'ANSWERED'
} as const

export type EmailFlagType = (typeof EmailFlagType)[keyof typeof EmailFlagType]


export const NotificationType = {
  NEW_LOGIN_DETECTED: 'NEW_LOGIN_DETECTED',
  LOW_MAILBOX_SPACE: 'LOW_MAILBOX_SPACE',
  PASSWORD_CHANGED: 'PASSWORD_CHANGED',
  WEEKLY_SECURITY_REPORT: 'WEEKLY_SECURITY_REPORT',
  NEW_EMAIL_RECEIVED: 'NEW_EMAIL_RECEIVED',
  PHISHING_DETECTED: 'PHISHING_DETECTED',
  MALWARE_DETECTED: 'MALWARE_DETECTED'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const SyncStatus = {
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED'
} as const

export type SyncStatus = (typeof SyncStatus)[keyof typeof SyncStatus]


export const AuditTargetType = {
  USER: 'USER',
  MAILBOX: 'MAILBOX',
  EMAIL: 'EMAIL',
  NOTIFICATION: 'NOTIFICATION'
} as const

export type AuditTargetType = (typeof AuditTargetType)[keyof typeof AuditTargetType]